Cleaning up stuff

+++ programs/steward/src/constants.rs
@@ -10,7 +10,7 @@ pub const EPOCH_PROGRESS_MAX: f64 = 0.99;
 // Cannot go more than 100 epochs without scoring
 pub const NUM_EPOCHS_BETWEEN_SCORING_MAX: u64 = 100;
 // Cannot score validators in under 100 slots, to submit 1 instruction per validator
-pub const COMPUTE_SCORE_SLOT_RANGE_MIN: usize = 100;
+pub const COMPUTE_SCORE_SLOT_RANGE_MIN: u64 = 100;
 #[cfg(feature = "mainnet-beta")]
 pub const VALIDATOR_HISTORY_FIRST_RELIABLE_EPOCH: usize = 520;
 #[cfg(not(feature = "mainnet-beta"))]
